seperate different search bars
each search bar is now it's own file
