+++ src/components/OfficerSelector.tsx
@@ -7,7 +7,7 @@ import {
   type FieldErrors,
 } from 'react-hook-form';
 import { type z } from 'zod';
-import { UserSearchBar } from './SearchBar';
+import { UserSearchBar } from './searchBar/UserSearchBar';
 
 type OfficerSelectorProps = {
   control: Control<z.infer<typeof createClubSchema>>;
